Adicionar "MiscUtils#hasInvite" para evitar code duplication
