+++ src/main/java/com/mrpowergamerbr/loritta/commands/vanilla/misc/SayCommand.kt
@@ -92,20 +92,8 @@ class SayCommand : AbstractCommand("say", listOf("falar"), CommandCategory.MISC)
 					whitelisted.add(it)
 				}
 
-				val matcher = Constants.URL_PATTERN.matcher(message)
-
-				while (matcher.find()) {
-					var url = matcher.group()
-					if (url.contains("discord") && url.contains("gg")) {
-						url = "discord.gg" + matcher.group(1).replace(".", "")
-					}
-
-					val inviteId = MiscUtils.getInviteId("http://$url") ?: MiscUtils.getInviteId("https://$url")
-
-					if (inviteId != null) { // INVITES DO DISCORD
-						if (inviteId != "attachments" && inviteId != "forums" && !whitelisted.contains(inviteId))
-							return // Tem convites válidos? Apenas ignore! A Lori irá aplicar as punições necessárias logo depois...
-					}
+				if (MiscUtils.hasInvite(message, whitelisted)) {
+					return
 				}
 			}
